requirements.txt
selenium
webdriver-manager
python-telegram-bot
python-dotenv
boto3